requirements.txt
appdirs==1.4.4
audioread==2.1.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
decorator==5.0.9
Flask==2.0.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
gcloud==0.18.3
googleapis-common-protos==1.53.0
greenlet==1.1.0
httplib2==0.19.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
jwcrypto==0.8
librosa==0.8.0
llvmlite==0.36.0
MarkupSafe==2.0.1
numba==0.53.1
numpy==1.20.3
oauth2client==4.1.3
opencv-python==4.5.2.52
packaging==20.9
pooch==1.3.0
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyparsing==2.4.7
Pyrebase4==4.4.3
python-jwt==3.3.0
requests==2.25.1
requests-toolbelt==0.9.1
resampy==0.2.2
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
SoundFile==0.10.3.post1
SQLAlchemy==1.4.15
threadpoolctl==2.1.0
urllib3==1.26.4
Werkzeug==2.0.1
WTForms==2.3.3
